+++ docs/src/news.md
@@ -4,6 +4,7 @@
 
 - News 
     - Functions **outsd**, **outod** and **outsdod**: Compute a SD, OD and SD-OD outlierness.
+    - Functions **occknn**, **occlknn**: One-class classification using kNN and local kNN distance-based outlierness.
 
 - Modifications
     - Code cleaning.