+++ docs/src/news.md
@@ -5,6 +5,7 @@
 - News 
     - Functions **outsd**, **outod** and **outsdod**: Compute a SD, OD and SD-OD outlierness.
     - Functions **occknn**, **occlknn**: One-class classification using kNN and local kNN distance-based outlierness.
+    - Functions **plotxyz**: 3D scatterplot.
 
 - Modifications
     - Code cleaning.